test: define a InMemoryConfiguration that can be shared
Each rest module was defining its own InMemoryConfiguration. It was a
pain when we introduce new InMemory implementation because we had to
add the declaration in 3 files.
Now there is only one InMemoryConfiguration located next to in memory
alternatives. This configuration is packaged in test artifact and can be
referenced in Spring configuration in each rest module.
